refactor(dashboard): extract fleet hero copy into its own module
Moving resolveFleetHeroCopy out of fleet-workspace.tsx keeps the
workspace under the structural debt ratchet after the connect-flow
flags were added.
